Better OIDC Support
Now the identity service requires a client secret, we need to keep it
secret by passing it in as a private environment variable.  This means
all authorization functionality needs to be moved server-side rather
than exposing it in an SPA.
