+++ charts/ui/Chart.yaml
@@ -4,8 +4,8 @@ description: A Helm chart for deploying Unikorn UI
 
 type: application
 
-version: v0.3.9-rc1
-appVersion: v0.3.9-rc1
+version: v0.3.9-rc2
+appVersion: v0.3.9-rc2
 
 icon: https://assets.unikorn-cloud.org/assets/images/logos/dark-on-light/icon.png
