+++ charts/ui/templates/deployment.yaml
@@ -30,7 +30,7 @@ spec:
         env:
         - name: PUBLIC_APPLICATION_VERSION
           value: {{ .Chart.Version }}
-        - name: PUBLIC_OAUTH2_ISSUER
+        - name: PUBLIC_IDENTITY_HOST
           value: https://{{ include "unikorn.identity.host" . }}
         - name: PUBLIC_REGION_HOST
           value: https://{{ include "unikorn.region.host" . }}
@@ -40,16 +40,31 @@ spec:
           value: https://{{ include "unikorn.application.host" . }}
         - name: PUBLIC_COMPUTE_HOST
           value: https://{{ include "unikorn.compute.host" . }}
-        - name: PUBLIC_OAUTH2_CLIENT_ID
+        - name: OIDC_CLIENT_ID
           {{- if .Values.oauth2.clientName }}
           value: {{ include "resource.id" .Values.oauth2.clientName }}
           {{- else }}
           value: {{ .Values.oauth2.clientID }}
           {{- end }}
-        - name: PUBLIC_OAUTH2_CLIENT_SECRET
+        - name: OIDC_CLIENT_SECRET
           value: {{ .Values.oauth2.clientSecret }}
+        {{- if .Values.tls.private }}
+        - name: NODE_EXTRA_CA_CERTS
+          value: /var/run/secrets/unikorn-cloud.org/ca.crt
+        {{- end }}
         securityContext:
           readOnlyRootFilesystem: true
+        {{- if .Values.tls.private }}
+        volumeMounts:
+        - name: unikorn-ui-ingress-tls
+          mountPath: /var/run/secrets/unikorn-cloud.org
+        {{- end }}
       serviceAccountName: unikorn-ui
       securityContext:
         runAsNonRoot: true
+      {{- if .Values.tls.private }}
+      volumes:
+      - name: unikorn-ui-ingress-tls
+        secret:
+          secretName: unikorn-ui-ingress-tls
+      {{- end }}
