+++ src/lib/oidc/index.ts
@@ -6,9 +6,7 @@ import type { JWTVerifyResult } from 'jose';
 import { env } from '$env/dynamic/public';
 
 // These are required variables from the environment.
-export const issuer = env.PUBLIC_OAUTH2_ISSUER || '';
-export const clientID = env.PUBLIC_OAUTH2_CLIENT_ID || '';
-export const clientSecret = env.PUBLIC_OAUTH2_CLIENT_SECRET || '';
+export const issuer = env.PUBLIC_IDENTITY_HOST || '';
 
 export type IDToken = {
 	// openid scope.
