+++ src/routes/oauth2/refresh/+page.svelte
@@ -1,0 +1,1 @@
+Nothing to see here.
